Move function 'create_aprs_login' to ogn.gateway
